Improve scraped results and add geolocation
There are two scripts, scrape.js and geolocate.js. Scrape will write the correctly formatted data to groups-seed.json and geolocate will read this file and output geolocated-groups-seed.json.

It’s likely that the geolocation script will hit network errors because of the number of items, so you can keep running it until the number of errors becomes constant (actual geocode errors, not network errors). You can do this because it diffs the already geolocated files against the scraped files and only geolocates the reminder.